task: filter products by category

+++ spa/src/services/category.js
@@ -2,23 +2,23 @@ import api from "./api";
 
 const categoryService = {
   create: (data) => {
-    return api.post("categories", data);
+    return api.post("admin/categories", data);
   },
   
   findAll: () => {
-    return api.get("categories");
+    return api.get("admin/categories");
   },
 
   findById: (id) => {
-    return api.get(`categories/${id}`);
+    return api.get(`admin/categories/${id}`);
   },
 
   categoryRemove: (id) => {
-    return api.delete(`categories/${id}`);
+    return api.delete(`admin/categories/${id}`);
   },
 
   categoryUpdate: (id, data) => {
-    return api.put(`categories/${id}`, data);
+    return api.put(`admin/categories/${id}`, data);
   },
 };
